Move stateless tree widget to a separate file
