+++ apps/test-app/frontend/src/components/tree-widget/TreeWidget.tsx
@@ -3,46 +3,24 @@
  * See LICENSE.md in the project root for license terms and full copyright notice.
  *--------------------------------------------------------------------------------------------*/
 
-import { forwardRef, useCallback, useEffect, useMemo, useRef, useState } from "react";
+import { forwardRef, useCallback, useEffect, useRef, useState } from "react";
 import { useResizeDetector } from "react-resize-detector";
-import { PropertyRecord } from "@itwin/appui-abstract";
-import {
-  ControlledTree,
-  DelayLoadedTreeNodeItem,
-  FilteringInput,
-  FilteringInputStatus,
-  SelectionMode,
-  TreeDataProvider,
-  TreeEventHandler,
-  TreeNodeItem,
-  TreeNodeRenderer,
-  TreeNodeRendererProps,
-  TreeRenderer,
-  useDebouncedAsyncValue,
-  useTreeModel,
-  useTreeModelSource,
-  useTreeNodeLoader,
-} from "@itwin/components-react";
+import { FilteringInput, FilteringInputStatus } from "@itwin/components-react";
 import { IModelApp, IModelConnection } from "@itwin/core-frontend";
-import { TreeNode, UnderlinedButton } from "@itwin/core-react";
-import { SchemaContext } from "@itwin/ecschema-metadata";
-import { ECSchemaRpcLocater } from "@itwin/ecschema-rpcinterface-common";
-import { Tab, Tabs, Text } from "@itwin/itwinui-react";
-import { DiagnosticsProps, InfoTreeNodeItemType, isPresentationInfoTreeNodeItem, PresentationInfoTreeNodeItem } from "@itwin/presentation-components";
-import { createECSqlQueryExecutor, createMetadataProvider } from "@itwin/presentation-core-interop";
-import { HierarchyNode, HierarchyProvider, IECSqlQueryExecutor, IMetadataProvider, RowsLimitExceededError } from "@itwin/presentation-hierarchy-builder";
-import { ModelsTreeDefinition } from "@itwin/presentation-models-tree";
+import { Tab, Tabs } from "@itwin/itwinui-react";
+import { DiagnosticsProps } from "@itwin/presentation-components";
 import { DiagnosticsSelector } from "../diagnostics-selector/DiagnosticsSelector";
+import { StatelessTreeWidget } from "./StatelessTreeWidget";
 import { Tree } from "./Tree";
 
-interface Props {
+export interface TreeWidgetProps {
   imodel: IModelConnection;
   rulesetId?: string;
   height?: number;
   width?: number;
 }
 
-export function TreeWidget(props: Omit<Props, "height" | "width">) {
+export function TreeWidget(props: Omit<TreeWidgetProps, "height" | "width">) {
   const [openTab, setOpenTab] = useState(0);
   const { width, height, ref } = useResizeDetector<HTMLDivElement>();
   const tabsClassName = "tree-widget-tabs";
@@ -77,7 +55,7 @@ export function TreeWidget(props: Omit<Props, "height" | "width">) {
   );
 }
 
-export function RulesDrivenTreeWidget(props: Props) {
+export function RulesDrivenTreeWidget(props: TreeWidgetProps) {
   const { rulesetId, imodel } = props;
   const [diagnosticsOptions, setDiagnosticsOptions] = useState<DiagnosticsProps>({ ruleDiagnostics: undefined, devDiagnostics: undefined });
   const [filter, setFilter] = useState("");
@@ -125,113 +103,7 @@ export function RulesDrivenTreeWidget(props: Props) {
   );
 }
 
-export function StatelessTreeWidget(props: Omit<Props, "rulesetId">) {
-  const [filter, setFilter] = useState("");
-  const [filteringStatus, setFilteringStatus] = useState(FilteringInputStatus.ReadyToFilter);
-  const [queryExecutor, setQueryExecutor] = useState<IECSqlQueryExecutor>();
-  const [metadataProvider, setMetadataProvider] = useState<IMetadataProvider>();
-  const [modelsTreeHierarchyProvider, setModelsTreeHierarchyProvider] = useState<HierarchyProvider>();
-  const [hierarchyLevelSizeLimit, setHierarchyLevelSizeLimit] = useState<{ [parentId: string]: number | "unbounded" | undefined }>({});
-  useEffect(() => {
-    const schemas = new SchemaContext();
-    schemas.addLocater(new ECSchemaRpcLocater(props.imodel.getRpcProps()));
-    setQueryExecutor(createECSqlQueryExecutor(props.imodel));
-    setMetadataProvider(createMetadataProvider(schemas));
-  }, [props.imodel]);
-
-  const { value } = useDebouncedAsyncValue(
-    useCallback(async () => {
-      if (metadataProvider && queryExecutor && filter !== "") {
-        setFilteringStatus(FilteringInputStatus.FilteringInProgress);
-        return ModelsTreeDefinition.createInstanceKeyPaths({ metadataProvider, queryExecutor, label: filter });
-      }
-      return [];
-    }, [metadataProvider, queryExecutor, filter]),
-  );
-
-  useEffect(() => {
-    if (metadataProvider && queryExecutor) {
-      const sharedProps = {
-        metadataProvider,
-        hierarchyDefinition: new ModelsTreeDefinition({ metadataProvider }),
-        queryExecutor,
-      };
-      if (value) {
-        setFilteringStatus(FilteringInputStatus.FilteringFinished);
-        setModelsTreeHierarchyProvider(
-          new HierarchyProvider({
-            ...sharedProps,
-            filtering: {
-              paths: value,
-            },
-          }),
-        );
-      } else {
-        setModelsTreeHierarchyProvider(new HierarchyProvider(sharedProps));
-      }
-    }
-  }, [queryExecutor, metadataProvider, value]);
-
-  const dataProvider = useMemo((): TreeDataProvider => {
-    return async (node?: TreeNodeItem): Promise<TreeNodeItem[]> => {
-      const parent: HierarchyNode | undefined = node ? (node as any).__internal : undefined;
-      const parentId = node?.id;
-      const limit = hierarchyLevelSizeLimit[parentId ?? ""];
-      try {
-        if (modelsTreeHierarchyProvider) {
-          return (await modelsTreeHierarchyProvider.getNodes({ parentNode: parent, hierarchyLevelSizeLimit: limit })).map(parseTreeNodeItem);
-        }
-        return [];
-      } catch (e) {
-        if (e instanceof RowsLimitExceededError) {
-          return [
-            createInfoNode(
-              node,
-              `${IModelApp.localization.getLocalizedString("Sample:controls.result-limit-exceeded")} ${limit!}`,
-              InfoTreeNodeItemType.ResultSetTooLarge,
-            ),
-          ];
-        }
-        // eslint-disable-next-line no-console
-        console.error(e);
-        return [];
-      }
-    };
-  }, [modelsTreeHierarchyProvider, hierarchyLevelSizeLimit]);
-  const modelSource = useTreeModelSource(dataProvider);
-  const nodeLoader = useTreeNodeLoader(dataProvider, modelSource);
-  const eventHandler = useMemo(() => new TreeEventHandler({ nodeLoader, modelSource }), [nodeLoader, modelSource]);
-  const treeModel = useTreeModel(modelSource);
-  const nodeRenderer = (nodeProps: TreeNodeRendererProps) => (
-    <StatelessTreeNodeRenderer
-      {...nodeProps}
-      onLimitReset={(parentId?: string) => setHierarchyLevelSizeLimit((map) => ({ ...map, [parentId ?? ""]: "unbounded" }))}
-    />
-  );
-  const { headerRef, treeHeight } = useTreeHeight(props.height);
-  return (
-    <>
-      <TreeWidgetHeader onFilterChange={setFilter} filteringStatus={filteringStatus} showFilteringInput={true} ref={headerRef} />
-      <div className="filtered-tree">
-        {treeHeight && props.width && (
-          <ControlledTree
-            model={treeModel}
-            eventsHandler={eventHandler}
-            nodeLoader={nodeLoader}
-            treeRenderer={(treeProps) => <TreeRenderer {...treeProps} nodeRenderer={nodeRenderer} />}
-            selectionMode={SelectionMode.Extended}
-            iconsEnabled={true}
-            width={props.width}
-            height={treeHeight}
-          />
-        )}
-        {filteringStatus === FilteringInputStatus.FilteringInProgress ? <div className="filtered-tree-overlay" /> : null}
-      </div>
-    </>
-  );
-}
-
-function useTreeHeight(height?: number) {
+export function useTreeHeight(height?: number) {
   const [treeHeight, setTreeHeight] = useState(0);
   const headerRef = useRef<HTMLDivElement>(null);
   useEffect(() => {
@@ -251,7 +123,7 @@ interface HeaderProps {
   onDiagnosticsOptionsChange?: (options: DiagnosticsProps) => void;
 }
 
-const TreeWidgetHeader = forwardRef(function TreeWidgetHeader(props: HeaderProps, ref: React.ForwardedRef<HTMLDivElement>) {
+export const TreeWidgetHeader = forwardRef(function TreeWidgetHeader(props: HeaderProps, ref: React.ForwardedRef<HTMLDivElement>) {
   const { onFilterChange, filteringStatus, showFilteringInput } = props;
   return (
     <div ref={ref} className="tree-widget-header">
@@ -281,68 +153,3 @@ const TreeWidgetHeader = forwardRef(function TreeWidgetHeader(props: HeaderProps
     </div>
   );
 });
-
-function parseTreeNodeItem(node: HierarchyNode): DelayLoadedTreeNodeItem {
-  if (node.children === undefined) {
-    throw new Error("Invalid node: children not determined");
-  }
-  return {
-    __internal: node,
-    id: JSON.stringify([...node.parentKeys, node.key]),
-    label: PropertyRecord.fromString(node.label, "Label"),
-    icon: node.extendedData?.imageId,
-    hasChildren: !!node.children,
-    autoExpand: node.autoExpand,
-  } as DelayLoadedTreeNodeItem;
-}
-
-function createInfoNode(parentNode: TreeNodeItem | undefined, message: string, type?: InfoTreeNodeItemType): PresentationInfoTreeNodeItem {
-  const id = `${parentNode ? parentNode.id : ""}/info-node/${message}`;
-  return {
-    id,
-    parentId: parentNode?.id,
-    label: PropertyRecord.fromString(message),
-    message,
-    isSelectionDisabled: true,
-    children: undefined,
-    type: type ?? InfoTreeNodeItemType.Unset,
-  };
-}
-
-interface StatelessTreeNodeRendererProps extends TreeNodeRendererProps {
-  onLimitReset: (parentId?: string) => void;
-}
-
-function StatelessTreeNodeRenderer(props: StatelessTreeNodeRendererProps) {
-  const nodeItem = props.node.item;
-  if (isPresentationInfoTreeNodeItem(nodeItem)) {
-    return (
-      <TreeNode
-        isLeaf={true}
-        label={
-          <span>
-            <Text isMuted>
-              <span>{nodeItem.message}</span>
-              {nodeItem.type === InfoTreeNodeItemType.ResultSetTooLarge && (
-                <span>
-                  <span> - </span>
-                  <UnderlinedButton
-                    onClick={() => {
-                      props.onLimitReset(nodeItem.parentId);
-                    }}
-                  >
-                    {`${IModelApp.localization.getLocalizedString("Sample:controls.remove-hierarchy-level-limit")}.`}
-                  </UnderlinedButton>
-                </span>
-              )}
-            </Text>
-          </span>
-        }
-        level={props.node.depth}
-        isHoverDisabled={true}
-      />
-    );
-  }
-
-  return <TreeNodeRenderer {...props} />;
-}
